package/wmenu: new package
wmenu is an efficient dynamic menu for Sway and wlroots based Wayland
compositors. It provides a Wayland-native dmenu replacement which
maintains the look and feel of dmenu.

Furthermore, as of commit [1], sway uses wmenu instead of dmenu-wayland
with the following commit message:

"""
Switch default config to wmenu-run
This removes the last dependency bit on dmenu. No need for
"swaymsg exec" anymore: wmenu-run handles the xdg-activation
shenanigans.
"""

As such, change sway/Config.in to select wmenu instead of dmenu-wayland when
a user selects swaybar.

[1] https://github.com/swaywm/sway/commit/b44015578a3d53cdd9436850202d4405696c1f52

Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>