board/bsh/imx8mn-bsh-smm-s2: erase the entire NAND chip
If an attempt is made to create a UBI volume and it already exists, the
operation fails. Therefore, before requesting the creation of a UBI
volume, we erase the entire NAND to ensure that no errors occur.

Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 97ac89eb115dbcd28efd1cf95d9a308f71b33041)
Signed-off-by: Thomas Perale <thomas.perale@mind.be>