package/cdrkit: add patch to fix CMake 4.0 build issue
CMake 4.0 requires to have a cmake_minimum_required() in
CMakeLists.txt, which cdrkit doesn't have, so ths commit adds a patch
adding the missing statement. We have chosen version 3.18 because that
the oldest version that we expect is 3.18. From
package/cmake/Config.in.host:

  # The minimum system cmake version we expect if 3.18 as provided by
  # Debian bullseye, that we use in our reference build docker image.

The patch cannot be upstreamed, as cdrkit basically no longer has any
upstream.

Fixes:

  https://autobuild.buildroot.org/results/3412e47836b54928a55c12b46549d6307ab623e7/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>