package/libical: fix deprecated CMake 4 policy
The commit [1] updated the version of libical to version 3.0.20 which
included a number of build fixes for newer CMake version.
This patch is not included in the 2025.02.x branch and libical remained
subject to a number of build error with CMake 4.

This patch backport the fix to remove the usage of the now deprecated
CMP0005 policy with CMake version 4 (see [2]).

[1] 8cdeeb536c package/libical: bump to version 3.0.20
[2] https://cmake.org/cmake/help/latest/policy/CMP0005.html

Signed-off-by: Thomas Perale <thomas.perale@mind.be>