package/samba4: bump version to 4.22.2
Release notes:
https://www.samba.org/samba/history/samba-4.22.2.html
https://www.samba.org/samba/history/samba-4.22.1.html
https://www.samba.org/samba/history/samba-4.22.0.html

Rebased patch 0001 due to upstream commit
https://github.com/samba-team/samba/commit/4dbbfcb00401d7a797154c812587bfe2dda0aea1

An import of the host-python ssl module was added to wscript with commit
https://github.com/samba-team/samba/commit/549f67a9e655149b1ede7e2858836cb7811cc78c
Select BR2_PACKAGE_HOST_PYTHON3_SSL to provide the new dependency.

Note: the 4.22.2 release note mention a security fix
for CVE-2025-0620. This commit is not marked as a security bump
because this CVE was addressed in a previous Buildroot commit
updating samba4 to version 4.21.6, which also contain this fix.

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
[Julien:
 - add a note explaining why this commit is not a security bump
 - add "select BR2_PACKAGE_HOST_PYTHON3" in Config.in
]
Signed-off-by: Julien Olivain <ju.o@free.fr>