package/python-glslang: bump to version 15.3.0
See changelog:
https://github.com/KhronosGroup/glslang/blob/15.3.0/CHANGES.md

Fixes build issue: error: ‘uint32_t’ has not been declared, encountered
on GCC v15, caused by cstdint not being implicitly included.

https://bugs.gentoo.org/937261

This commit also adds host-spriv-tools in _DEPENDENCIES to use the
version packaged in Buildroot. It also adds
ALLOW_EXTERNAL_SPIRV_TOOLS=ON in _CONF_OPTS, which was added in upstream
commit [1], introduced in version 13.0.0.

Also, when trying to build this python-glslang version 15.3.0 on a host
providing cmake < 3.26, the configuration fails with the error message:

    CMake Error at CMakeLists.txt:345 (add_test):
      Error evaluating generator expression:

        $<TARGET_RUNTIME_DLL_DIRS:spirv-remap>

      Expression did not evaluate to a known generator expression

This happens in the Buildroot docker reference image, which provides a
cmake version 3.25. To workaround this issue which happen in glslang
testing code, this commit disable those tests by adding
GLSLANG_TESTS=OFF in _CONF_OPTS.

[1] https://github.com/KhronosGroup/glslang/commit/3805888a57c07904b6bbdec639bfe5f078e26342

Signed-off-by: Joseph Kogut <joseph.kogut@gmail.com>
[Julien:
 - resolve merge conflict
 - update changelog link to use version tag
 - disable glslang tests
]
Signed-off-by: Julien Olivain <ju.o@free.fr>